make camera moving speed independent of refresh rate

+++ include/guik/camera/arcball_camera_control.hpp
@@ -16,12 +16,12 @@ public:
   virtual void reset_center() override;
   virtual void lookat(const Eigen::Vector3f& pt) override;
 
-  virtual void mouse(const Eigen::Vector2i& p, int button, bool down) override;
-  virtual void drag(const Eigen::Vector2i& p, int button) override;
+  virtual void mouse(const Eigen::Vector2f& p, int button, bool down) override;
+  virtual void drag(const Eigen::Vector2f& p, int button) override;
   virtual void scroll(const Eigen::Vector2f& rel) override;
 
-  virtual void updown(int p) override;
-  virtual void arrow(const Eigen::Vector2i& p) override;
+  virtual void updown(double p) override;
+  virtual void arrow(const Eigen::Vector2f& p) override;
 
   virtual Eigen::Vector2f depth_range() const override;
 
@@ -45,8 +45,8 @@ protected:
 
   bool left_button_down;
   bool middle_button_down;
-  Eigen::Vector2i mouse_down_pos;
-  Eigen::Vector2i drag_last_pos;
+  Eigen::Vector2f mouse_down_pos;
+  Eigen::Vector2f drag_last_pos;
 };
 
 }  // namespace guik
